+++ FirebaseCombineSwift/Sources/Auth/MultiFactorResolver+Combine.swift
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#if canImport(Combine) && swift(>=5.0) && canImport(FirebaseAuth)
+#if canImport(Combine) && swift(>=5.0)
 
   #if os(iOS) || targetEnvironment(macCatalyst)
 
@@ -47,4 +47,4 @@
 
   #endif // os(iOS) || targetEnvironment(macCatalyst)
 
-#endif // canImport(Combine) && swift(>=5.0) && canImport(FirebaseAuth)
+#endif // canImport(Combine) && swift(>=5.0)