+++ Firestore/core/src/core/transaction_runner.h
@@ -67,7 +67,7 @@ class TransactionRunner
   core::TransactionUpdateCallback update_callback_;
   core::TransactionResultCallback result_callback_;
   remote::ExponentialBackoff backoff_;
-  int retries_left_;
+  int attempts_remaining_;
 };
 
 }  // namespace core