+++ scripts/setup_quickstart.sh
@@ -30,6 +30,15 @@ SAMPLE=$1
 
 RELEASE_TESTING=${2-}
 
+WORKSPACE_DIR="quickstart-ios/${SAMPLE}"
+PODFILE="quickstart-ios/"$SAMPLE"/Podfile"
+
+if [[ ! -z "${LEGACY:-}" ]]; then
+  WORKSPACE_DIR="quickstart-ios/${SAMPLE}/Legacy${SAMPLE}Quickstart"
+  PODFILE="quickstart-ios/"$SAMPLE"/Legacy${SAMPLE}Quickstart/Podfile"
+fi
+
+
 # Installations is the only quickstart that doesn't need a real
 # GoogleService-Info.plist for its tests.
 if check_secrets || [[ ${SAMPLE} == "installations" ]]; then
@@ -39,20 +48,20 @@ if check_secrets || [[ ${SAMPLE} == "installations" ]]; then
   export FIREBASE_POD_REPO_FOR_DEV_POD=`pwd`
 
   git clone https://github.com/firebase/quickstart-ios.git
-  $scripts_dir/localize_podfile.swift quickstart-ios/"$SAMPLE"/Podfile "$RELEASE_TESTING"
+  $scripts_dir/localize_podfile.swift "$WORKSPACE_DIR"/Podfile "$RELEASE_TESTING"
   if [ "$RELEASE_TESTING" == "nightly_release_testing" ]; then
     set +x
-    sed -i "" '1i\'$'\n'"source 'https://${BOT_TOKEN}@github.com/FirebasePrivate/SpecsTesting.git'"$'\n' quickstart-ios/"$SAMPLE"/Podfile
+    sed -i "" '1i\'$'\n'"source 'https://${BOT_TOKEN}@github.com/FirebasePrivate/SpecsTesting.git'"$'\n' "$PODFILE"
     set -x
     echo "Source of Podfile for nightly release testing is updated."
   fi
   if [ "$RELEASE_TESTING" == "prerelease_testing" ]; then
     set +x
-    sed -i "" '1i\'$'\n'"source 'https://${BOT_TOKEN}@github.com/FirebasePrivate/SpecsReleasing.git'"$'\n' quickstart-ios/"$SAMPLE"/Podfile
+    sed -i "" '1i\'$'\n'"source 'https://${BOT_TOKEN}@github.com/FirebasePrivate/SpecsReleasing.git'"$'\n' "$PODFILE"
     set -x
     echo "Source of Podfile for prerelease testing is updated."
   fi
-  cd quickstart-ios/"$SAMPLE"
+  cd "${WORKSPACE_DIR}"
 
   # To test a branch, uncomment the following line
   # git checkout {BRANCH_NAME}
@@ -65,7 +74,11 @@ if check_secrets || [[ ${SAMPLE} == "installations" ]]; then
     bundle exec pod install
   fi
 
+  if [[ ! -z "${LEGACY:-}" ]]; then
+    cd ..
+  fi
+
   # Add GoogleService-Info.plist to Xcode project
-  ruby ../scripts/info_script.rb "${SAMPLE}"
+  ruby ../scripts/info_script.rb "${SAMPLE}" "${LEGACY:-}"
   cd -
 fi