+++ FirebaseAppCheck/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Unreleased
+- [fixed] Bump Promises dependency. (#8365)
+# v8.3.0 -- M99
 - [added] Token API for 3P use. (#8266)
-# 8.2.0 -- M98
+# v8.2.0 -- M98
 - [added] Apple's App Attest attestation provider support. (#8133)
 - [changed] Token auto-refresh optimizations. (#8232)
 # v8.0.0 -- M95