+++ FirebaseAppDistribution/CHANGELOG.md
@@ -1,4 +1,4 @@
-# Unreleased
+# v8.3.0-beta
 - [changed] Sign out the Tester when the call to fetch releases fails with an unauthorized error (#8270).
 
 # v7.3.0-beta