+++ FirebaseCombineSwift/CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
 - [feature] Added Combine support for Cloud Functions for Firebase
-- [feature] WIP: Added Combine support for Firebase Auth
-- [feature] WIP: Added Combine support for Firebase Storage
+- [feature] Added Combine support for Firebase Auth
+- [feature] Added Combine support for Firebase Storage
+- [feature] Added Combine support for Cloud Firestore